chore: Renamed App to AppProviders
