+++ src/index.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./App";
+import App from "./AppProviders";
 import { BrowserRouter as Router } from "react-router-dom";
 
-ReactDOM.render(<App />, document.getElementById("root")
-);
+ReactDOM.render(<AppProviders />, document.getElementById("root"));
